requirements.txt
emoji
python-graphql-client==0.3.0
Pillow
